+++ src/helpers/window-factory.js
