+++ src/presentation/html/index.html
@@ -3,7 +3,7 @@
 <head>
   <meta charset="UTF-8">
   <title>Spotify - now playing</title>
-  <link href="css/index.css" rel="stylesheet" type="text/css"/>
+  <link href="../css/index.css" rel="stylesheet" type="text/css"/>
   <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.5.0/css/solid.css" integrity="sha384-rdyFrfAIC05c5ph7BKz3l5NG5yEottvO/DQ0dCrwD8gzeQDjYBHNr1ucUpQuljos" crossorigin="anonymous">
   <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.5.0/css/fontawesome.css" integrity="sha384-u5J7JghGz0qUrmEsWzBQkfvc8nK3fUT7DCaQzNQ+q4oEXhGSx+P2OqjWsfIRB8QT" crossorigin="anonymous">
 </head>
@@ -11,6 +11,6 @@
   <div id="player-container"></div>
   <div id="loader-container" class="loader"></div>
   <div id="add-container"></div>
-  <script src="js/main-view.js"></script>
+  <script src="../js/main-view.js"></script>
 </body>
 </html>
